locales

+++ src/views/hr/employee/history-record.vue
@@ -2,20 +2,20 @@
   <div class="page">
     <div class="page__header">
       <span class="page__header--title pointer" @click="$router.go(-1)">
-        <i class="iconfont icon-retreat"></i>历史上传记录
+        <i class="iconfont icon-retreat"></i>{{ $t('history_record_page_title') }}
       </span>
       <div class="page__header--btn">
         <el-dropdown class="mr30" trigger="click">
           <span class="el-dropdown-link">
-            导出<i class="el-icon-arrow-down el-icon--right"></i>
+            {{ $t('common_export') }}<i class="el-icon-arrow-down el-icon--right"></i>
           </span>
           <el-dropdown-menu slot="dropdown">
-            <el-dropdown-item @click.native="handleHeaderExportExcel(1)">导出全部</el-dropdown-item>
-            <el-dropdown-item @click.native="handleHeaderExportExcel(2)">导出成功记录</el-dropdown-item>
-            <el-dropdown-item @click.native="handleHeaderExportExcel(3)">导出失败记录</el-dropdown-item>
+            <el-dropdown-item @click.native="handleHeaderExportExcel(1)">{{ $t('history_record_header_btn1_1') }}</el-dropdown-item>
+            <el-dropdown-item @click.native="handleHeaderExportExcel(2)">{{ $t('history_record_header_btn1_2') }}</el-dropdown-item>
+            <el-dropdown-item @click.native="handleHeaderExportExcel(3)">{{ $t('history_record_header_btn1_3') }}</el-dropdown-item>
           </el-dropdown-menu>
         </el-dropdown>
-        <el-button size="small" @click="handleBatchDelete()">批量删除</el-button>
+        <el-button size="small" @click="handleBatchDelete()">{{ $t('history_record_header_btn2') }}</el-button>
       </div>
     </div>
 
@@ -27,14 +27,14 @@
         :inline="true"
         :model="searchParams"
         @keyup.enter.native="handleQuery('firstPage')">
-        <el-form-item label="文件名">
+        <el-form-item :label="$t('history_record_search_label1')">
           <el-input
             v-model="searchParams.name"
             @blur="searchParams.name = searchParams.name.trim()"
             maxlength="32"
             clearable></el-input>
         </el-form-item>
-        <el-form-item label="上传时间">
+        <el-form-item :label="$t('history_record_search_label2')">
           <date-picker-range
             :startTime.sync="searchParams.begTime"
             :endTime.sync="searchParams.endTime"
@@ -43,18 +43,18 @@
         </el-form-item>
 
         <el-form-item class="btn__search">
-          <el-button icon="iconfont icon-search" @click="handleQuery('firstPage')">查询</el-button>
+          <el-button icon="iconfont icon-search" @click="handleQuery('firstPage')">{{ $t('common_search') }}</el-button>
         </el-form-item>
       </el-form>
 
       <el-table :data="historyList" height="100%" @selection-change="handleSelectionChange">
         <el-table-column type="selection" width="55"></el-table-column>
-        <el-table-column prop="createTime" label="时间" min-width="100"></el-table-column>
-        <el-table-column prop="name" label="文件名" min-width="120"></el-table-column>
-        <el-table-column prop="susNum" label="成功人数" min-width="50"></el-table-column>
-        <el-table-column prop="failNum" label="失败人数" min-width="50"></el-table-column>
-        <el-table-column prop="totalNum" label="注册照数量" min-width="50"></el-table-column>
-        <el-table-column label="操作" min-width="100" class-name="table__icon--btn">
+        <el-table-column prop="createTime" :label="$t('history_record_table_column1')" min-width="100"></el-table-column>
+        <el-table-column prop="name" :label="$t('history_record_table_column2')" min-width="120"></el-table-column>
+        <el-table-column prop="susNum" :label="$t('history_record_table_column3')" min-width="50"></el-table-column>
+        <el-table-column prop="failNum" :label="$t('history_record_table_column4')" min-width="50"></el-table-column>
+        <el-table-column prop="totalNum" :label="$t('history_record_table_column5')" min-width="50"></el-table-column>
+        <el-table-column :label="$t('history_record_table_column6')" min-width="100" class-name="table__icon--btn">
           <template slot-scope="scope">
             <el-button
               class="mr10"
@@ -63,19 +63,19 @@
                 name: 'exportDetail',
                 query: {id: scope.row.id, time: scope.row.createTime}
               })">
-              查看详情
+              {{ $t('history_record_table_btn1') }}
             </el-button>
             <el-dropdown class="mr10" trigger="click">
               <span class="el-dropdown-link f-blue pointer">
-                导出<i class="el-icon-arrow-down el-icon--right"></i>
+                {{ $t('common_export') }}<i class="el-icon-arrow-down el-icon--right"></i>
               </span>
               <el-dropdown-menu slot="dropdown">
-                <el-dropdown-item @click.native="handleExportExcel(1, scope.row.id)">导出全部</el-dropdown-item>
-                <el-dropdown-item @click.native="handleExportExcel(2, scope.row.id)">导出成功记录</el-dropdown-item>
-                <el-dropdown-item @click.native="handleExportExcel(3, scope.row.id)">导出失败记录</el-dropdown-item>
+                <el-dropdown-item @click.native="handleExportExcel(1, scope.row.id)">{{ $t('history_record_header_btn1_1') }}</el-dropdown-item>
+                <el-dropdown-item @click.native="handleExportExcel(2, scope.row.id)">{{ $t('history_record_header_btn1_2') }}</el-dropdown-item>
+                <el-dropdown-item @click.native="handleExportExcel(3, scope.row.id)">{{ $t('history_record_header_btn1_3') }}</el-dropdown-item>
               </el-dropdown-menu>
             </el-dropdown>
-            <el-tooltip effect="dark" content="删除" placement="top" :enterable="false">
+            <el-tooltip effect="dark" :content="$t('common_delete')" placement="top" :enterable="false">
               <i class="iconfont icon-delete" @click="handleDelete(scope.row)"></i>
             </el-tooltip>
           </template>
@@ -137,12 +137,12 @@ export default {
     },
     handleHeaderExportExcel (type) {
       if (!this.checkedList.length) {
-        this.$message.warning('请先选择需要导出的历史记录！')
+        this.$message.warning(this.$t('history_record_warning1'))
         return
       }
 
-      this.$confirm('批量导出选中的历史记录，系统将自动为您生成excel文件', '导出', {
-        confirmButtonText: '导出',
+      this.$confirm(this.$t('history_record_confirm1'), this.$t('common_export'), {
+        confirmButtonText: this.$t('common_export'),
         center: true
       }).then(() => {
         this.checkedList.forEach((ele, index) => {
@@ -154,7 +154,7 @@ export default {
             const a = document.createElement('a')
             const blob = new Blob([res], { type: 'application/vnd.ms-excel' })
             a.href = URL.createObjectURL(blob)
-            a.setAttribute('download', '人员导出列表.xlsx')
+            a.setAttribute('download', this.$t('history_record_excel_name'))
             document.body.appendChild(a)
             a.click()
             URL.revokeObjectURL(a.href)
@@ -166,29 +166,42 @@ export default {
       window.open(`${this.baseUrl}/employee/exportHistory?type=${type}&recordId=${id}`)
     },
     handleDelete ({ id }) {
-      api.deleteHistory(id).then(res => {
-        if (res.success) {
-          this.$message.success('删除成功！')
-          this.handleQuery()
-        }
-      })
+      this.$confirm(this.$t('history_record_confirm2'), this.$t('common_delete'), {
+        customClass: 'delete__box--confirm',
+        confirmButtonText: this.$t('common_delete'),
+        center: true
+      }).then(() => {
+        api.deleteHistory(id).then(res => {
+          if (res.success) {
+            this.$message.success(this.$t('history_record_success1'))
+            this.handleQuery()
+          }
+        })
+      }, () => {})
     },
     handleBatchDelete () {
       if (!this.checkedList.length) {
-        this.$message.warning('请先选择需要删除的历史记录！')
+        this.$message.warning(this.$t('history_record_warning3'))
         return
       }
-      let params = ''
-      this.checkedList.forEach(ele => {
-        params += `&recordIds=${ele.id}`
-      })
-      params = params.substr(1)
-      api.batchDeleteHistory(params).then(res => {
-        if (res.success) {
-          this.$message.success('删除成功！')
-          this.handleQuery()
-        }
-      })
+
+      this.$confirm(this.$t('history_record_warning2'), this.$t('common_delete'), {
+        customClass: 'delete__box--confirm',
+        confirmButtonText: this.$t('common_delete'),
+        center: true
+      }).then(() => {
+        let params = ''
+        this.checkedList.forEach(ele => {
+          params += `&recordIds=${ele.id}`
+        })
+        params = params.substr(1)
+        api.batchDeleteHistory(params).then(res => {
+          if (res.success) {
+            this.$message.success(this.$t('history_record_success1'))
+            this.handleQuery()
+          }
+        })
+      }, () => {})
     }
   }
 }
